--- a/@vates/nbd-client/multi.mjs
+++ b/@vates/nbd-client/multi.mjs
@@ -41,14 +41,16 @@ export default class MultiNbdClient {
     }
     if (connectedClients.length < this.#clients.length) {
       warn(
-        `incomplete connection by multi Nbd, only ${connectedClients.length} over ${this.#clients.length} expected clients`
+        `incomplete connection by multi Nbd, only ${connectedClients.length} over ${
+          this.#clients.length
+        } expected clients`
       )
       this.#clients = connectedClients
     }
   }
 
   async disconnect() {
-    asyncEach(this.#clients, client => client.disconnect(), {
+    await asyncEach(this.#clients, client => client.disconnect(), {
       stopOnError: false,
     })
   }