--- a/packages/xo-server/src/api/backup-ng.mjs
+++ b/packages/xo-server/src/api/backup-ng.mjs
@@ -22,13 +22,13 @@ const SCHEMA_SETTINGS = {
           minimum: 0,
           optional: true,
         },
-        preferNbd: {
-          type: 'boolean',
-          optional: true,
-        },
         nbdConcurrency: {
           type: 'number',
           minimum: 1,
+          optional: true,
+        },
+        preferNbd: {
+          type: 'boolean',
           optional: true,
         },
       },