### Description

review by commit, do not squash 

### homelab
adding latency through tc to xenbr0 in dom0 , single host, not limited by write on remote

|   |   0ms | 10ms   | 20ms  |50ms   |
|---|---|---|---|---|
|  Nbd |  260 MB/s|  42MB/S |  23 MB/S|  11 MB/S |
|  8xNbd |  260 MB/S|  230 MB/s | 145MB/s  | 80 MB/s |

### xo-lab 
real world latency from the datacenter to my home . bandwidth limit should be around~400Mbps = 50MB/s

|   | 20ms, ethernet | 25-35ms wifi |
|---|---|---|
|Nbd|  21MB/s  |  8MB/s  |
| 8x Nbd | 42MB/s |  15MB/s  |

  
  latency 20 , single, then multiple link, then latency 50 single then multiple
![image](https://github.com/vatesfr/xen-orchestra/assets/50174/ce36058d-bee2-42dd-bf78-1a28cb837a1f)


Ajout de l'ui : 
![image](https://github.com/vatesfr/xen-orchestra/assets/50174/80bd94c9-5581-40ad-939a-8098168d5d4f)

### Checklist

- Commit
  - Title follows [commit conventions](https://bit.ly/commit-conventions)
  - Reference the relevant issue (`Fixes #007`, `See xoa-support#42`, `See https://...`)
  - If bug fix, add `Introduced by`
- Changelog
  - If visible by XOA users, add changelog entry
  - Update "Packages to release" in `CHANGELOG.unreleased.md`
- PR
  - If UI changes, add screenshots
  - If not finished or not tested, open as _Draft_