@fbeauchamp 2023-12-05
tested and didn't improve performance: 
* asking for smaller blocks instead of 2MB (64kb=> 1MB)
* asking for bigger block when reading contigous data ( up to 20MB)
* replacing readchunk (which add and remove listenner on the fly by one listenner : little CPU decrease
* increasing readahead ( no gain if more than 4 on a singl link ) 
*