@fbeauchamp 2023-12-08
tested : backup a few VM with and without NBD multi , compare bit to bit the raw disk : no difference